+++ packages/neovim/src/attach.ts
@@ -1,1 +1,0 @@
-export { attach } from './attach/attach';