+++ packages/neovim/src/utils/transport.test.ts
@@ -1,0 +1,30 @@
+/* eslint-env jest */
+import { EventEmitter } from 'node:events';
+import { Readable, Writable } from 'node:stream';
+import * as msgpack from '@msgpack/msgpack';
+import { attach } from '../attach/attach';
+import { exportsForTesting } from './transport';
+
+describe('transport', () => {
+  it('throws on invalid RPC message', done => {
+    const invalidPayload = { bogus: 'nonsense' };
+    const onTransportFail: EventEmitter = exportsForTesting.onTransportFail;
+    onTransportFail.on('fail', (errMsg: string) => {
+      expect(errMsg).toEqual(
+        "invalid msgpack-RPC message: expected array, got: { bogus: 'nonsense' }"
+      );
+      done();
+    });
+
+    // Create fake reader/writer and send a (broken) message.
+    const fakeReader = new Readable({ read() {} });
+    const fakeWriter = new Writable({ write() {} });
+
+    const nvim = attach({ reader: fakeReader, writer: fakeWriter });
+    void nvim; // eslint-disable-line no-void
+
+    // Simulate an invalid message on the channel.
+    const msg = msgpack.encode(invalidPayload);
+    fakeReader.push(Buffer.from(msg.buffer, msg.byteOffset, msg.byteLength));
+  });
+});